Remove "Pass" check level
There is no valid reason for a check to be defined, only for the results to be ignored. This was used as a default behavior when no level was configured for the check, but that indicates a configuration error and should not be ignored. So I now error exit when this situation occurs.