Use more appropriate variable name for project type filter
The term "project type" is used both to refer to the type of the project filter, which may be "all", and to the specific
project type (e.g., sketch, library, platform). This can result in some ambiguity in the code. The use of "filter" in
variable names that are used exclusively for project type filters may make the code easier to follow.