Generalize JSON schema handling code
Previously, the JSON schema handling code was all under the libraryproperties package, and written specifically for use
in validating library.properties against the JSON schema. However, there will be be a need to work with schemas in other
contexts, most immediately for writitng the schema tests, but also likely when additional schemas are added for the
other Arduino project configuration data files.

So it will be helpful to have the general purpose schema handling code
in a dedicated package, leaving only the library.properties-specific aspects of the code in the libraryproperties
package.