Add permissive and strict library.properties JSON schemas
I started the arduino-check project with the simplistic vision of it being a tool used to enforce the Arduino
Sketch/Library/Platform/Package Index Specifications. However this approach neglects some important use cases:

- Library Manager index management. Specification violations don't always result in breakage, so library authors may
resent a strict requirement that all libraries added to the index be specification-compliant. More significantly, large
numbers of non-compliant libraries are already in the index, so a change to blocking non-specification compliant
releases would likely be very disruptive. For this reason, a "permissive" subset of the specification is needed.

- Advocating best practices. Although not mandated by the specification, some practices result in an improved experience
for the user, and thus should be advocated by Arduino. It would be unreasonable to enforce these practices on unwilling
users, but they can be advocated for with warnings and by offering willing users an option to use the tool in a
"pedantic" mode. For this reason, a "strict" superset of the specification is needed.

JSON schemas offer a standardized method of defining a data format. This form of information can be easily used by
consumers other than the arduino-check tool. So I think it's the best approach to define these compliance levels in
schemas, rather than directly in Go code.

Significant components of the schemas are identical. To avoid the need to maintain duplicate content, it's important to
use referencing. While working on this, I found that it it was confusing to work with a separate file for the definition
of each compliance level of each document property, especially when considering how changing one schema affects all the
schemas that reference it, and the schemas that reference those schemas... I settled on the approach of using a
monolithic definitions schema, resulting in the parent schema being nearly identical collection of references to that
file, which are unlikely to need to be updated at all in the future, even when modifications are made to the
definitions schema.

With the idea that schemas for the other Arduino project configuration files will be created in the future to expand the
scope of arduino-check, I put the definitions that can be used by other schemas in a dedicated definitions schema, from
which they can be referenced by additional schemas.