Add full schema pointer matching capability to schema.ValidationErrorMatch()
When a schema validation error is related to one of the logic inversion keywords ("not", "oneOf"), the schema pointer
path provided in the schema validation package's validation results data does not extend past that keyword.

Capability has now been added for the error to be matched against the full JSON pointers under the problematic keywords.

In combination with structuring the schemas in a manner that puts the logic inversion keywords at the lowest possible
level, this makes it possible to fully determine the cause of a validation failure.