Change `--permissive` flag to `--compliance`
The initial design of the tool was based on the idea of simply enforcing the Arduino project specifications. It soon
became clear that, due to the legacy of not strictly enforcing specification compliance on Library Manager libraries
where the non-compliance doesn't result in a significant problem, there was a need for a "permissive" mode. In addition,
the need was found for a "strict" mode, in which the tool will advocate best practices above and beyond the
specifications.

For this purpose, the `--permissive` flag was added. However, along the way the original concept of enforcing the
specifications was lost. Permissive mode is an unfortunate necessity. Strict mode is great, but may not be to some
people's taste. So a configuration mode where the tool simply check specification compliance remains useful, and is a
reasonable default.