Remove "All" check mode
This mode was an initial unsuccessful attempt at achieving what was later done by the "Default" check mode.

The "All" check mode is no longer necessary. It can have a harmful effect when used in place of "Default" because it
makes the behavior of the check configuration dependent on the order the check configuration fields are processed in,
which is confusing and not guarantee to remain the same. The "Default" check mode is always overridden by an explicit
check mode match, so it's behavior is well defined and intuitive.