Search child folders of subprojects parent folder for subprojects
Previously, the subprojects parent folder was searched for subprojects. Due to the recursive search setting, that still
found library example sketches (after unnecessarily checking the examples folder itself for a sketch). However, the
subproject search of platform bundled libraries is not recursive (because the libraries must be directly under the
libraries folder), meaning these subprojects were not discovered.