Use properties.SafeLoadFromPath() to load library.properties
In the event of a problem loading, this will return an empty map rather than nil. The error already serves as the
indicator of this problem for the invalid library.properties check, so having an empty map as the output only simplifies
the handling of that situation in the other checks.