Provide reason for not running checks
Return an explanatory message when it is impossible to run a check due to the required information being unavailable.
Previously, the infrastructure was in place to display this output to the user, but there was no output to display, so
you just got a confusing "notice:" output with no notice to follow.