Suppress "notice:" output on check not run when there is no explanation
The check functions may output an explanation when they are unable to run, which is printed to the console following a
"notice:" check result label. If no explanation was provided by the check function author, this resulted in a confusing
line in the console. Although it is best for an explanation to always be provided, this is not guaranteed to always be
done, so it's more user friendly to suppress the check level display in this situation.