Return effective check level from checklevel.CheckLevel()
Previously, the level returned was for a failure result, regardless of the actual result of the check, with the
effective check level adjusted according to the check result in the result package. The code is more maintainable and
easier to understand if the check level is determined in a single location. The checklevel package is the logical place
for that.