Add checks for missing license file
If a file with one of the names supported by the licensee gem is in the root of the repository containing the text of a
standard license, GitHub recognizes the license type and enables some nice integrations. So having such a file is best
practices.

These checks are only for the presence of the file, rather than their contents.