Allow empty depends field
The regular expression was only intended to check for prohibited characters, but it was inadvertently enforcing a
minimum length of 1, which was also inherited by the name property schema (which already has a separate minimum length
schema). There is currently no plans to add a minimum length requirement for the depends field, and if that is desired,
it is better accomplished via the "minLength" schema.