Add --verbose flag
Previously, the tool alway provides output for every check it runs. This is not necessarily a bad thing because in
`--format text` mode it provides a progress indicator, and in `--format json` mode it provides information on all the
checks that are run under the current tool configuration.

However, as the number of checks grows, the amount of output has become a bit overwhelming and may make it hard to find
the information for the failed checks. So it makes sense to default to only showing information on the failed checks
while providing the option to show all checks.