Make --recursive flag a string input
Previously, the `--recursive` flag had a Boolean type. That is logical, since it uses values of true and false. However,
the fact that the default value of this flag is true changes matters. With the other Boolean flags, the default value
is false, so when you want to set it to true you only need to use the flag without a value, as is standard. But with a
default true flag it's necessary to unset the value, which can't be done with a bare flag, so you must pass false to the
flag. You would expect to be able to do this with either a space separator or an equals sign, just like
with the string flags, but the former is not supported:

$ ./arduino-check --recursive false
Configuration error: PROJECT_PATH argument false does not exist

Only the equals sign is supported (`--recursive=false`).

So I think the UI is more intuitive with this flag a string type, so the behavior is consistent with the other flags.