Use appropriate check result value when determined not relevant
There are various extents to which a check can be run:

1. Skipped due to check configuration (e.g., check is configured to run for libraries but the project is a sketch).
2. Not run due to a problem not directly related to the check (e.g., invalid library.properties data format).
3. Skipped because the check function determined it was not applicable (e.g., formating check on undefined optional library.properties field).
4. Run.

Each has an associated check result value: Skip, NotRun, Pass/Fail.

Previously, the "NotRun" result was used for both (2) and (3), but "Skip" is much more appropriate for (3) because this
is a situation where there was no reason to run the check, whereas "NotRun" indicates that the check should have been
run, but couldn't.