Reorder check configurations
During the course of development, the order of the check configurations in the array has become quite random. The checks
are run in order of the configurations, so this is exposed to the user, and also makes the code a bit less easy to
follow.

The check functions and their tests have reordered to match the check configurations.