Remove unwanted Python-Markdown extensions
Python-Markdown extensions are used to provide additional capabilities to the documentation website generation. Because
Python-Markdown is extremely faithful to the original Markdown specification, despite the fact that the world has moved
on to a more capable implementation, some extensions are essential to provide the Markdown capability the documentation
editors will expect.

However, some additional extenstions were added to the Arduino CLI MkDocs configuration which
provide syntax not supported by any Markdown implementation, and which are not needed.

If we later find the pressing
need for such capability, we can add in the necessary extensions. Until then, having these extensions could result in
unexpected behavior.