Disable ajv-cli strict mode when validating CI workflows
The JSON schema for GitHub Actions workflow files from the JSON Schema Store is not passing metaschema validation with
the new release of ajv-cli. It seems the tool now defaults to a "strict" mode and that schema (note: it's an external
resource, not one of arduino-lint's schemas) is not compliant.