Only require official prefix on standalone library names
In the case of official platform bundled libraries, it doesn't make sense to req	uire the Arduino_ prefix since
equivalent platform bundled libraries are often provided by each platform, both official and 3rd party. The "namespace"
provided by the prefix is only needed in the context of standalone libraries. In the case of platform bundled libraries,
the platform can provide the branding.