Make all packages internal
The Go packages of this project were written solely for the use within the project. Making them internal avoids them
unintentionally becoming part of the public API, resulting in a greatly increased potential maintenance and support
burden and limitations on further development work. It would be interesting to identify the components of the internal
API that might be useful to others and expose them intentionally as they reach a state of relative stability. However,
for now the public API should be considered only the command line interface and the JSON report structure.