Remove bashisms from install script
Unfortunately, `local` is Bash-specific. The script's shebang (#!/bin/sh) indicates POSIX compliance was intended, so
the use of `local` is not appropriate. I used the function name as a prefix "namespace" for the previously local
variables in the attempt to achieve somewhat of the same effect in a POSIX-compliant manner.