Move installation script to etc subfolder
I would prefer to avoid cluttering up the root of the repository too much. The reason is that when people visit the
repository on GitHub, they must scroll down past the root directory listing before getting to the readme. Some files and
folders must be in the root, but there are some others (e.g., gon.config.hcl) that can be moved to a subfolder, so this
move can be part of a larger campaign.