Improve handling of missing package index
Following the convention of the other project types, the user can specify the path to a folder containing a package
index.

Previously, when no file resembling a package index was found in that folder, the linting process would error. It is
correct for this to be considered an error, however the panic process was used for this purpose. Although it printed a
friendly error message at the start of the output, a panic also outputs a traceback. That traceback is very useful for
debugging the code of the action, but it's not very user friendly. The established convention is that panics are used
for errors that should not occur if the tool is operating correctly, which errors caused by the user input are handled
only with a friendly error message.

The new behavior is to use the rule system to handle this situation. This allows for the addition of additional rules
that might provide the user with information about the cause of the error.