Show brief after rule ID in verbose text output
The rule brief is a short description of the rule. When arduino-lint is run in `--verbose` mode, the ID of each rule is
displayed. While this serves to show the progress of linting each project, the user will not know the significance of
these IDs, and previously wouldn't get any other context unless the rule was violated (in which case the failure message
is displayed), so it is somewhat cryptic  Displaying the brief after the rule ID explains the rules that are being run
without increasing the volume of output significantly.