Use optional plurality in prohibited character rule messages
The rule covers one or more prohibited characters, so use of a hard singular/plural in the rule message might lead to
confusion. At the same time, the alternative (s) uses more characters, so is less well suited to the rule brief. The
compromise is to use the singular "prohibited character" in the brief text, and the flexible "prohibited character(s)"
in the less length-constrained rule message, which will always be displayed if the rule is violated.