Make schema.Validate() general purpose
In order to work with the nested data structures of the platform configuration files, it's necessary to modify this
function. Although the necessary properties.Map conversion could be done by schema.Validate(), this will not support
working with the package index. So it's best to make it support all requirements now, rather than having to modify it
yet again.