Unify properties to map conversion code
In order to do validation against the JSON schema the native properties.Map format of the Arduino project configuration
files must be converted to the map[string]interface{} type. This process is made more complex by the need to selectively
recurse the data into multiple key levels.

Previously, this was done in three different places in the code, each
tailored to the specific needs of the configuration files it worked with, and using slightly different approaches.

Unifying all that code into a single general purpose function will make the code easier to maintain.