Allow incomplete hidden board definitions
Board definitions using the hide property may have never been intended to provide a usable board definition. An example
of this is defining only the minimum necessary properties to associate a board name with a VID/PID pair.

Since these boards are not visible to the user, there is not the expectation that it must support compiling and
uploading. So some of the properties which are required for other boards don't apply to these.