Use most appropriate test assertion types
The require package's assertion functions immediately terminate the test when the assertion fails. The assert package's
assertion functions allow the test to continue when the assertion fails.

For this reason, the require functions should always and only be used when the failure of the assertion indicates that
the results of all subsequent tests will be meaningless.

In cases where the failure of the assertion has no bearing on the results of subsequent tests, the assert functions
should always be used so the test results can be as comprehensive as possible to allow all issues to be fixed before the
next test run.