Temporarily disable platform documentation rules
Although the official boards platforms are located in the root of the repository, some 3rd party platforms have the platform under the architecture subfolder of the repository in order to be friendly to manual installation (Boards Manager installation automatically generates this folder).

The problem is that the license and readme should be in the repo root, but the project path is not the repository root when this alternative repository structure is in use. So this results in a false positive from the readme and license checks.

It's only a warning when not in strict mode, but having a spurious warning is annoying and also blocks people from being able to use strict mode.

These are important rules and I plan to find a solution to allow this rule to be applied to platforms of both structures once the development resources are available to make this happen. But since the rule is currently not compatible with all valid use cases, the best immediate action is to disable it.