Refactor subproject discovery code
Subproject discovery must be done under any circumstances. Previously, there was duplicate code to do this for both file project paths and folder project paths, resulting in the code being more difficult to maintain.

In order to accomodate this change, the error message in the event of no projects being discovered was reworded to be applicable to either project path type.