Use major version ref of action in workflow
At the time the workflow was written the authors of the rtCamp/action-slack-notify GitHub Action did not provide a major version ref. This meant that it was necessary to pin the action to a specific version.

Since then, a few new patch releases have been made, meaning an outdated version of the action was in use as a consequence of the pinning. 

The action now offers a `v2` major ref. Use of this ref will cause the workflow to benefit from ongoing development to the action up until such time as a new major release is made, at which time we would need to evaluate whether any changes to the workflow are required by the breaking change that triggered the major release before updating the major ref (e.g., `v3`).