Add a Library Manager configuration for linting while indexing releases
Some of Arduino Lint's rules are based on the data in the published Arduino Library Manager index. These allow library
maintainers to check compliance with the Library Manager submission or update requirements.

Work is now in progress to use Arduino Lint to validate library releases during the generation of the index. In this
usage, data from the published index is not relevant, meaning the results from those rules are inaccurate and so
downloading the index for the sake of them is a waste. So another configuration mode is needed to to allow the
index-based rules to be disabled.

Since this mode is of no use to normal users of Arduino Lint, it is enabled via an environment variable so that no
unnecessary complexity is added to the standard UI of the tool. The environment variable name is
`ARDUINO_LINT_LIBRARY_MANAGER_INDEXING`.