Add min length rule for library.properties url field
Previously, the permissive compliance mode only required that the library.properties url field was present. The field was
allowed to be empty in this mode.

Investigation revealed that the Arduino Library Manager indexer system, which Arduino Lint's permissive compliance mode
is intended to match, requires the url field value to be at least one character long:

https://github.com/arduino/libraries-repository-engine/blob/c395aaf14e7568f585a6a96350b584682a3c7feb/libraries/metadata/metadata.go#L110