Add rule for 3rd party library.properties maintainer using "Arduino"
3rd party libraries are not maintained by Arduino. Even when 3rd party libraries are based on official libraries, it is
best practice for the library.properties `maintainer` field to be updated to reflect the true maintainer of the project.

There is already rule LP027 to promote accuracy in documenting maintainership by prohibiting maintainer values that start
with "Arduino". However, this might not cover all inaccurate maintainer declarations. For this reason, a new rule (LP057)
is added here to prohibit maintainer values in 3rd party libraries from containing the term "Arduino" anywhere
(case insensitive). Violations of this rule only result in an error when in "strict" compliance mode.