Use Arduino CLI infrastructure for querying library index
In the past, I have found that using Arduino CLI infrastructure for handling projects is not always the right choice
because Arduino CLI is intended to be resilient to sub-ideal project configurations when feasible, whereas Arduino Lint
is intended to enforce or encourage ideal project configurations.

However, Arduino Lint is not intended to lint the Library Manager index, so using Arduino CLI's code for querying the
index should be safe and avoid the problems associated with the code duplication that would be the alternative.