Add support for dependency constraints to LP048
The library.properties depends field is used to specify the library dependencies. Although currently undocumented,
Arduino CLI and the Library Manager indexer both have provisional support for dependency version constraints.

Previously, LP048 did not have any support for the constraint syntax, which caused a spurious failure when they were in
use. This commit adds support for the constraint syntax and also causes the constraint information to be used when
checking if the dependency is provided by the index.