Add convenience function for "format" schema result parsing
The JSON schema validation result parsing infrastructure provides general purpose capabilities for identifying the
instance and schema pointers related to a validation failure. However, when a specific type of schema keyword is used in
multiple places, it is convenient to create a dedicated wrapper function to facilitate that use case. Previously the
"format" keyword was only used for one thing in the library.properties schemas, which made it not worth a convenience
function. The package index schema uses "format" keyword also, so it now becomes worthwhile having a convenience parsing
function for it.