Add convenience function for "type" schema result parsing
The JSON schema validation result parsing infrastructure provides general purpose capabilities for identifying the
instance and schema pointers related to a validation failure. However, when a specific type of schema keyword is used in
multiple places, it is convenient to create a dedicated wrapper function to facilitate that use case. Although present in
the schemas for the sake of completeness, there are no rules for checking schema validation results for the "type"
keyword because the Arduino "properties" data format treats all values as strings. For this reason, there was no convenience
function for this keyword. Due to using the JSON data format, which allows for any type, there will be type checking
schema-based rules for the package index project type. So it now becomes worthwhile having a convenience parsing function
for the "type" keyword.