Add convenience function for "additionalProperties" schema result parsing
The JSON schema validation result parsing infrastructure provides general purpose capabilities for identifying the
instance and schema pointers related to a validation failure. However, when a specific type of schema keyword is used in
multiple places, it is convenient to create a dedicated wrapper function to facilitate that use case. Previously the
"additionalProperties" keyword was not used in any rules, which made it not worth a convenience function. The package
index schema uses the "additionalProperties" keyword, so it now becomes worthwhile having a
convenience parsing function for it.