Allow full JSON pointers in schema validation result parsing functions
When used with the "properties" data format, and most especially with the flat format of the library.properties file that
was an initial focus of the project's development, JSON pointers are not a terribly intuitive way to refer to a data path.

In this context, it seemed sensible to pre-fill the root prefix on the instance pointer argument of the JSON schema
validation result parsing functions. However, now that those functions are being used with the JSON formatted package
index data file, the requirement to omit this prefix from the JSON pointers is very confusing.

Since the argument is already a regex, it's simple to allow both styles of instance data path references.