Make package index loading function handle missing file
Each possible problem with the project is handled by a dedicated rule. One such problem is that the project was not
found. Each rule must be configured so that unrelated problems with the project will not cause them to break or provide
incorrect results. This can result in a lot of redundant boilerplate code. Before running a rule against the package
index data, it is always necessary to check whether the data is available. If not, running the rule is abandoned with a
"NotRun" result. This which results in just such boilerplate, but is likely to be unavoidable without breaking from the
modular architecture of the rules system.

If the package index file itself was not found, then the loading of the data can not be accomplished, and thus a missing
package index should also be cause for abandoning the run of data rules. The check for data can serve as an implicit
check for the existence of the data source, meaning additional rule function boilerplate for file existence check can be
avoided.