Provide natively loaded data as package index project data
Arduino CLI's package index loading function can can be useful in its current capacity as a general check on the validity
of the package index as far as Arduino CLI is concerned, used for that purpose by a single dedicated rule.

However, it can not be used as a data source for other rules because Arduino CLI should attempt to be resilient to bad
data when possible, while Arduino Lint needs to detect the very bad data that Arduino CLI might correct.

So the data from the native loading code must be used in all rules, while the value from Arduino CLI's load is solely in
whether or not it returned an error.