Use JSON pointer as package component ID when standard ID component empty
In cases where the data used to assemble the standard ID used to refer to a package component is undefined, the JSON
pointer is used.

Previously, when the data is defined, but empty, the standard ID format was used. But this will not be very meaningful.
In this case, the JSON pointer will serve as a more useful ID.