Provide object-level packages[].platforms[].boards[] package index data to rules
Data checking rules for package indexes need to iterate over its object-level components. To facilitate this, slices of
such objects accompanied by the necessary metadata are generated as package index project data.

This was already done for packages, platforms, and tools, but not for boards. Although board objects are trivial, this
seems like it will provide for cleaner rule functions that apply to to these objects than the alternative of generating
the data on a per-rule basis.