Rescind policy that commits should start with "[skip changelog]"
The raw release changelog is generated from the commit messages. The `arduino/create-changelog` action is configured to
omit any commit message that is prefixed with "[skip changelog]".

After extensive use of this approach in the Arduino CLI repository, my conclusion is that it provides a miniscule time
savings for the release manager at the cost of massive pollution of the valuable commit history and a significantly less
contributor friendly repository. For this reason, I have never employed this practice in my work on Arduino Lint and I do
not want contributors to do so either.