Move rule output control logic to result package
Since output always has had the same format regardless of the tool configuration, with the configuration only determining
whether or not it will be printed, it made sense to place the printing control logic at the point of printing, always
outputting the message from the result package.

It has now been determined that there are different formatting requirements for verbose and non-verbose output. This
means that it makes more sense to move that logic to the result package and simply return an empty string when no output
should be printed. This refactoring in preparation for the formatting changes to the non-verbose output.