Simplify non-verbose rule output
Previously, the output resulting from rule violations was always prefaced with the rule ID and result:

Rule SD001 result: fail

This is necessary in verbose output mode because in that mode the result of every rule applied to the project is printed.
This is the only way to identify the rule and result when a rule passes.

The situation is different non-verbose mode, where output is only printed when a rule is violated. This means that:

- The rule is identified by the rule message
- The rule result is implicit

So this preface doesn't really serve any purpose in non-verbose mode. It is somewhat cryptic, so may make the output less
approachable to the users. Because the rule ID is still sometimes of value as a succinct and unequivocal way to refer to
a specific rule, it is printed in a less prominent location at the end of the rule message.