Use ">>" in place of "-" between package index ID components
Since package index data occurs in arrays, the rules must indicate to the user the specific element(s) in violation. In
the case of standard components (e.g. platform), the form of the IDs is well established (e.g., `arduino:avr@1.8.3`).
However, that I know of, there is no standard form for unequivocally identifying a package index data subcomponent (e.g.,
a tool dependency of a specific platform release).

I arbitrarily chose a dash to combine the IDs of the two components (e.g.,
`arduino:avr-gcc@7.3.0-atmel3.6.1-arduino7 - x86_64-linux-gnu`). I received feedback that this was not very clear, with a
suggestion to use ">>" instead of the "-" (e.g., `arduino:avr-gcc@7.3.0-atmel3.6.1-arduino7 >> x86_64-linux-gnu`).