Put rule ID suffix on new line when rule message is multi-line
Readability of rule message output may be improved by inserting strategic line breaks. In this case, the current behavior
of appending the rule ID suffix to the non-verbose rule violation output on the same line will not always be appropriate.
So when the rule message contains explicit line breaks, the rule ID should be appended on a new line, with the previous
behavior preserved in the case of single line rule messages (which will be wrapped according to the column size).