Use strategic newlines to improve rule output readability
In the cases where a project may violate a rule multiple times, the rule message contains a list of identifiers for the
violations. Previously, this always took the form of a comma-separated list. These identifiers can sometimes be quite
long. When automatic line wrapping occurred in the middle of an identifier, it could make the message difficult to
interpret, especially now that Arduino Lint wraps its own output in a tabular form.

The solution is to insert line breaks into the rule message where appropriate. This is done for all violation identifier
lists that are likely to have long identifiers, as well as additional locations as needed.

The rule ID is appended to the rule message and there are also situations where the rule message structure will contain
other text after the violation identifier list. For this reason, the list is indented to visually differentiate it from
the non-list message text that follows the list.