Run "Check Certificates" workflow on pull requests when appropriate
It facilitates review to run a workflow when pull request is submitted proposing changes to it. Because this workflow
requires access to secrets which are not available when a pull request is submitted from a fork, the workflow was not
previously configured to support this. However, it is most likely that pull requests for changes to this particular
workflow will come from people who have write permissions in the repository, and who often submit PRs from branches in
the repository itself. In this case, the workflow does have access to the secrets.