Add `repository_dispatch` event trigger to "Check Certificates" workflow
This event allows triggering workflows via the GitHub API. This might be useful for triggering an immediate check in
multiple relevant repositories after an external change, or some automated process. Although we don't have any specific
need for this event at the moment, the event has no impact on the workflow, so there is no reason against having it. It
is the sort of thing that can end up being useful if it is already in consistently in place, but not worth setting up on
demand, since the effort to set it up is greater than the effort to trigger all the workflows manually.