Use `go fmt` instead of `gofmt` for formatting Go code
Since `go fmt` is only a formatter, it was necessary to use `gofmt` for the formatting check task, since that command has
an option to skip writing the code. In order to ensure consistency between the two, I also used `gofmt` for the
formatting task. The unfortunate thing about gofmt is that it works from a list of paths, rather than a list of packages
as is the case with `go vet`, `go fix`, and `golint`.

With the removal of the superfluous formatting check task, the motivation for using gofmt in the formatting task is
removed, freeing me to use the superior `go fmt`.