Move Go formatting check to "Check Go" workflow
GitHub Actions provides a paths filter feature that allows you to restrict workflow run triggers to changes to specific
paths within the repository. This can greatly improve the efficiency of the CI system by avoiding running irrelevant
checks. The effective use of this feature requires that the purpose of a workflow be focused on a specific type of file
or project component. The "Check Formatting" workflow was the antithesis of that; grouping formatting checks of all types
of files.